moved Navbar component to separate folder
